+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change challenge computation adding the public key to the hash [#3]
+
 ## [0.1.0] - 2024-01-08
 
 ### Added
@@ -14,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add initial commit, this package continues the development of [dusk-schnorr](https://github.com/dusk-network/schnorr/) at version `0.18.0` under the new name: jubjub-schnorr
 
 <!-- ISSUES -->
+[#3]: https://github.com/dusk-network/jubjub-schnorr/issues/3
 
 <!-- VERSIONS -->
 [Unreleased]: https://github.com/dusk-network/jubjub-schnorr/compare/v0.1.0...HEAD